+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bundleup/core
 
+## 0.0.5
+
+### Patch Changes
+
+- fix: using sdk to create session
+
 ## 0.0.4
 
 ### Patch Changes