+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bundleup/nextjs
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bundleup/core@0.0.5
+  - @bundleup/react@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes