+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundleup/react
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bundleup/core@0.0.5
+
 ## 0.0.6
 
 ### Patch Changes